adjust app, auto-test register

+++ backend/contracts/tool/contract.ps1
@@ -4,15 +4,23 @@ Write-Output "license that can be found in the license file."
 Set-Location $PSScriptRoot
 Set-Location ..
 
-Write-Output "|-> * npm install prepared. "
-Write-Output ""
-Start-Sleep -Milliseconds 1000
+Write-Output "|-> * Next step is npm 'install', if you are already install successfully one time, you can skip it.  "
 
-npm install zeppelin-solidity
-npm install -g truffle@4.1.14
+Start-Sleep -Milliseconds 300
 
-Write-Output "|-> * npm install finished. "
-Write-Output ""
+$confirm = Read-Host "|-> * Make sure you read the tip above, do you want to install? [Y/N] "
+
+if ($confirm -eq "Y") { # no matter if the character is big-case or small-case
+    Write-Output "|-> * npm install prepared. "
+    Write-Output ""
+    Start-Sleep -Milliseconds 1000
+
+    npm install zeppelin-solidity
+    npm install -g truffle@4.1.14
+
+    Write-Output "|-> * npm install finished. "
+    Write-Output ""
+}
 
 truffle version
 Write-Output "|-> * truffle migrate prepared. "
@@ -24,5 +32,3 @@ truffle migrate --network geth
 Stop-Transcript 
 Write-Output "|-> * truffle migrate finished. "
 Write-Output "|-> * End. "
-
-Start-Sleep -Milliseconds 5000
